Use newGVOP to construct GVOPs
Ideally, the type of OP that backs GVOPs should be irellevant. newGVOP will work
no matter if GVOPs are really PADOPs or SVOPs.

In a perfect world we'd probably use the same code for both with and without
ithreads, and decide what blass to bless the wrapper for the resulting OP* into
based on op->op_type. Though that'd require some refactoring for how B:: objects
are constructed, and is therefore not part of this patch.